Revert "Features and improvements"
