--- .gitignore
@@ -1,4 +1,0 @@
-*.txt
-venv/*
-*cpython*
-__pycache__
