--- variables/codes.py
@@ -1,10 +1,0 @@
-# Errors
-status_error = 0
-status_ok = 1
-
-# Requests
-request_successful = 2
-request_error = 3
-ignore = 4
-
-unkdownError = 666
